general bug fixes, unit testing for custom kernels, sanity checking that error decreases and results are deterministic across runs, some differences b/t cublas and ATLAS discovered, not sure what is causing that, but it was all of cublas (not just batched), so probably OK

+++ tpsopt/tps.py
@@ -53,7 +53,7 @@ def tps_kernel_matrix2(x_na, y_ma):
     return tps_apply_kernel(distmat, dim)
 
 def tps_eval(x_ma, lin_ag, trans_g, w_ng, x_na):
-    K_mn = tps_kernel_matrix2(x_ma, x_na)
+    K_mn = tps_kernel_matrix2(x_ma, x_na).astype(np.float32)
     return np.dot(K_mn, w_ng) + np.dot(x_ma, lin_ag) + trans_g[None,:]
 
 def tps_grad(x_ma, lin_ag, _trans_g, w_ng, x_na):
